+++ labb6/src/labb6/supermarketSimulator/PayTime.java
@@ -25,7 +25,7 @@ public class PayTime {
      * @param currentTime den nuvarande tiden i state
      * @return absoluta tiden som händelsen kommer att hända
      */
-    public double calculate(int currentTime) {
+    public double calculate(double currentTime) {
         return currentTime + rand.next();
     }
